Split iOS local database tests
